pyproject.toml
[tool.poetry]
name = "pipelines"
version = "0.1.0"
description = ""
authors = ["Gabriel Gazola Milan <gabriel.gazola@poli.ufrj.br>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
dbt-bigquery = "^1.6.1"
google-cloud-storage = "^2.10.0"
prefect = "1.4.1"
prefeitura-rio = { git = "https://github.com/prefeitura-rio/prefeitura-rio.git", rev = "e1fb218837d3231e07be90d71a07ec1c9d9593f5", branch = "feat/reimplement-templates", extras = [
    "pipelines",
    "pipelines-templates",
] }
python-dotenv = "^1.0.0"
azure-storage-blob = "^12.19.0"
pandas = "^2.1.4"
gitpython = "^3.1.40"
pendulum = "^3.0.0"
sqlalchemy = "^2.0.25"
mysqlclient = "2.2.3"
validate-docbr = "^1.10.0"
cryptography = "42.0.4"
gspread="^5.12.4"
unidecode = "^1.3.8"
ping3 = "^4.0.5"
psycopg2-binary = "^2.9.9"
discord-py = "^2.3.2"
seaborn = "^0.13.2"
selenium = "4.16.0"
pydrive2 = "^1.19.0"
httpx = "^0.27.0"
simpledbf = "^0.2.6"
tables = "^3.9.2"
pyreaddbc = "^1.2.0"
openpyxl = "^3.1.5"


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
flake8 = "^6.1.0"
pre-commit = "^3.3.3"
taskipy = "^1.12.0"
isort = "^5.12.0"


[tool.poetry.group.ci]
optional = true

[tool.poetry.group.ci.dependencies]
networkx = "^3.1"
loguru = "^0.7.0"
typer = "^0.9.0"

[tool.black]
line-length = 100
target-version = ["py310"]
include = "\\.pyi?$"

[tool.isort]
profile = "black"

[tool.taskipy.tasks]
lint = "black . && isort . && flake8 ."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"